Fixed two bugs: - After removing abs from calculating absolute number of lanes to changes, I was subtracting from an unsinged integer, which resulted in a overflow - I now only inhibit lane changes, when the cars speed is dropping below 10mph and not as soon as the speed is reduced.
This is a stable version which worked on my local simulation.
Running the same code on the web simulation resulted in entirely
different results due to much shorter simulation cycles.

On my local machine it took the simulator/code 7 cycles to finish
an iteration, in the web environment simulation only 3.

Calculated Accelerations and Speeds are very different fron the
two simulations.